requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
lpips==0.1.4
MarkupSafe==2.1.4
matplotlib==3.8.2
mpld3==0.5.10
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
scipy==1.12.0
six==1.16.0
sympy==1.12
torch==1.9.1+cu113
torchvision==0.10.1+cu113
torchaudio==0.9.1

#torchaudio==2.1.2
#torchvision==0.16.2
tqdm==4.66.2
triton==2.1.0
typing_extensions==4.9.0
urllib3==2.1.0
zipp==3.17.0
